Modify type of `context`

+++ fleac/flea_ast.hpp
@@ -13,7 +13,7 @@ class BaseAST {
 public:
   virtual ~BaseAST() = default;
   virtual void print(std::ostream &out) const = 0;
-  virtual int64_t const_eval(SymbolTable *stb, uint32_t context = 0) = 0;
+  virtual int64_t const_eval(SymbolTable *stb, uint64_t context = 0) = 0;
   friend std::ostream &operator<<(std::ostream &out, const BaseAST &ast);
 };
 
@@ -26,7 +26,7 @@ public:
   SymbolTable stb;
   std::vector<std::unique_ptr<BaseAST>> func_def_l;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
   void insertFunc(BaseAST *func_def);
 };
 
@@ -42,7 +42,7 @@ public:
   std::unique_ptr<BaseAST> block;
   std::unique_ptr<std::vector<std::unique_ptr<BaseAST>>> fparam_l;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // FuncFParam ::= BType IDENT;
@@ -53,7 +53,7 @@ public:
   char type_id;
   std::unique_ptr<std::string> ident;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // Block ::= "{" {BlockItem} "}";
@@ -62,7 +62,7 @@ public:
   BlockAST(std::vector<std::unique_ptr<BaseAST>> *item_l) : item_l(item_l) {}
   std::unique_ptr<std::vector<std::unique_ptr<BaseAST>>> item_l;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // BlockItem ::= Decl | Stmt;
@@ -87,7 +87,7 @@ public:
   char type_id;
   std::unique_ptr<std::vector<std::unique_ptr<BaseAST>>> def_l;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // VarDef ::= IDENT "=" InitVal;
@@ -100,7 +100,7 @@ public:
   std::unique_ptr<std::string> ident;
   std::unique_ptr<BaseAST> init_val;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // InitVal ::= Exp;
@@ -112,7 +112,7 @@ public:
   bool is_const;
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // class StmtAST : public BaseAST {
@@ -129,7 +129,7 @@ public:
   ExpStmtAST(BaseAST *exp) : exp(exp) {}
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // RetStmt ::= "return" Exp ";";
@@ -138,21 +138,21 @@ public:
   RetStmtAST(BaseAST *exp) : exp(exp) {}
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 class BreakStmtAST : public BaseAST {
 public:
   BreakStmtAST() {}
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 class ContinueStmtAST : public BaseAST {
 public:
   ContinueStmtAST() {}
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // AssignStmt ::= LVal "=" Exp ";";
@@ -162,7 +162,7 @@ public:
   std::unique_ptr<BaseAST> lval;
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 class IfStmtAST : public BaseAST {
@@ -174,7 +174,7 @@ public:
   std::unique_ptr<BaseAST> then_stmt;
   std::unique_ptr<BaseAST> else_stmt;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 class WhileStmtAST : public BaseAST {
@@ -184,7 +184,7 @@ public:
   std::unique_ptr<BaseAST> cond;
   std::unique_ptr<BaseAST> stmt;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // Exp ::= EqExp;
@@ -195,7 +195,7 @@ public:
   bool is_const;
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // PrimaryExp ::= "(" Exp ")" | LVal | Number;
@@ -204,7 +204,7 @@ public:
   PrimaryExpAST(BaseAST *exp) : exp(exp) {}
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // LVal ::= IDENT;
@@ -213,7 +213,7 @@ public:
   LValAST(std::string *ident) : ident(ident) {}
   std::unique_ptr<std::string> ident;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
   bool is_const(SymbolTable *stb) const;
 };
 
@@ -223,7 +223,7 @@ public:
   NumberAST(int32_t number) : number(number) {}
   int32_t number;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // UnaryOp ::= "+" | "-" | "!";
@@ -234,7 +234,7 @@ public:
   char unary_op;
   std::unique_ptr<BaseAST> exp;
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 class BinExpAST : public BaseAST {
@@ -248,7 +248,7 @@ public:
 
   BinExpAST(char op, BaseAST *lhs, BaseAST *rhs) : op(op), lhs(lhs), rhs(rhs) {}
   void print(std::ostream &out) const override final;
-  int64_t const_eval(SymbolTable *stb, uint32_t context = 0) override final;
+  int64_t const_eval(SymbolTable *stb, uint64_t context = 0) override final;
 };
 
 // MulExp ::= UnaryExp | (MulExp ("*" | "/" | "%") UnaryExp);
